requirements.txt
Django
djangorestframework==3.3.3
gunicorn==19.6.0
